Re-unify module and interface in ConcreteModuleType
We made `ModuleInfo` a union initially to represent the idea that a
submodule could either be a regular module or a module interface.

This PR represents module interfaces as a ConcreteModuleType with no
info (e.g.  no "sugaredness"), and with the interface type as the
underlying `jitType_`. This has the effect of reducing the special
casing around adding/maintaining module info.

ghstack-source-id: 915b3fe0c81027194d9b2c56860f913cbf408461
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29825